Animation Object
Animation Component available.
added Image Components as well.
made the Texture as separated class.

+++ Bullet.cpp
@@ -6,8 +6,8 @@ Bullet::Bullet(Vector2 position, const char* directory)
 {
     m_transform.w = s_kWidth;
     m_transform.h = s_kHeight;
-    m_transform.x = m_position.m_x;
-    m_transform.y = m_position.m_y;
+    m_transform.x = (int)m_position.m_x;
+    m_transform.y = (int)m_position.m_y;
 
 }
